Replace.Megaparsec.Internal.ByteString

Description

This internal module is for ByteString specializations.

The functions in this module are intended to be chosen automatically by rewrite rules in the Replace.Megaparsec module, so you should never need to import this module.

Names in this module may change without a major version increment.

Parser combinator

sepCapByteString

Arguments

:: forall e s m a. (MonadParsec e s m, s ~ ByteString)

=> m a
The pattern matching parser sep

-> m [Either (Tokens s) a]

anyTillByteString

Arguments

:: forall e s m a. (MonadParsec e s m, s ~ ByteString)

=> m a
The pattern matching parser sep

-> m (Tokens s, a)